requirements.txt
aiohttp
websockets
#numpy
simplejson
requests
python-dotenv